+++ backends.hpp
@@ -27,7 +27,7 @@
 #define MFEM_BLOCK_ID(k) 0
 #define MFEM_THREAD_ID(k) 0
 #define MFEM_THREAD_SIZE(k) 1
-#define MFEM_FOREACH_THREAD(i,k,N) for(int i=0; i<N; i++)
+#define MFEM_FOREACH_THREAD(i,k,N) for(int64_t i=0; i<N; i++)
 
 template <typename T>
 T AtomicAdd(T &add, const T val)
